refactor: 사용자 아바타 공통 컴포넌트
